Disable testTimeoutPerConnection when security enabled
This test doesn't work with security enabled because (since #90534) we
require the mock socket to complete the TLS handshake. It only works on
BSD/MacOS anyway, so I think it's fine to skip it when using TLS.

Closes #93613
